requirements.txt
asgiref==3.3.1
astroid==2.4.2
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-tempus-dominus==5.1.2.13
django-widget-tweaks==1.4.8
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
wrapt==1.12.1
